+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/decorators/CrateLicenseDecorator.kt
@@ -13,7 +13,6 @@ import software.amazon.smithy.rust.codegen.core.smithy.CodegenContext
 import software.amazon.smithy.rust.codegen.core.smithy.RustCrate
 
 class CrateLicenseDecorator : RustCodegenDecorator<ClientProtocolGenerator, ClientCodegenContext> {
-    override val name: String = "CrateLicense"
     override val order: Byte = 0
 
     override fun extras(codegenContext: ClientCodegenContext, rustCrate: RustCrate) {
