+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/decorators/HttpResponseChecksumDecorator.kt
@@ -30,7 +30,6 @@ private fun HttpChecksumTrait.requestValidationModeMember(
 }
 
 class HttpResponseChecksumDecorator : RustCodegenDecorator<ClientProtocolGenerator, ClientCodegenContext> {
-    override val name: String = "HttpResponseChecksum"
     override val order: Byte = 0
 
     override fun operationCustomizations(
