+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/servicedecorators/route53/TrimResourceId.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.rustsdk.servicecustomizations.route53
+package software.amazon.smithy.rustsdk.servicedecorators.route53
 
 import software.amazon.smithy.model.node.Node
 import software.amazon.smithy.model.shapes.ShapeId
